server/requirements.txt
Flask>=0.10.1
leancloud-sdk>=1.0.9
flask-cors>=2.0